Extract Users::Authentication module from User model
I started to do this when `User` was too long due to the multi-line association declarations.  I ended up making them single line, but this still seemed like a nice change.
